scraper/requirements.txt
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
pdfplumber==0.11.4
apscheduler==3.10.4
python-dotenv==1.0.1
selenium==4.27.1
webdriver-manager==4.0.2
